:truck: 크롤러의 경로를 themeModules 내부로 이동
