+++ backend/tsconfig.json
@@ -28,11 +28,11 @@
       "@group/*": ["./src/modules/userModules/group/*"],
       "@chat/*": ["./src/modules/userModules/chat/*"],
       "@brand/*": ["./src/modules/themeModules/brand/*"],
-      "@theme/*": ["./src/modules/themeModules/theme/*"],
       "@branch/*": ["./src/modules/themeModules/branch/*"],
+      "@crawlerUtils/*": ["./src/modules/themeModules/crawlerUtils/*"],
+      "@theme/*": ["./src/modules/themeModules/theme/*"],
       "@diary/*": ["./src/modules/diaryModules/diary/*"],
       "@comment/*": ["./src/modules/diaryModules/comment/*"],
-      "@crawler/*": ["./src/crawler/*"]
     }
   }
 }
